# docs/swagger.yaml
definitions:
...
  cmd_api-gateway.forgotPasswordResp:
    properties:
      message:
        type: string
    type: object
...
  cmd_api-gateway.loginResp:
    properties:
      token:
        $ref: '#/definitions/github_com_r0lm0_go-saas-api_internal_auth.TokenPair'
      user:
        $ref: '#/definitions/github_com_r0lm0_go-saas-api_internal_auth.User'
    type: object
  cmd_api-gateway.logoutResp:
    properties:
      message:
        type: string
    type: object
...
  cmd_api-gateway.refreshResp:
    properties:
      token:
        $ref: '#/definitions/github_com_r0lm0_go-saas-api_internal_auth.TokenPair'
      user:
        $ref: '#/definitions/github_com_r0lm0_go-saas-api_internal_auth.User'
    type: object
  cmd_api-gateway.registerResp:
    properties:
      user:
        $ref: '#/definitions/github_com_r0lm0_go-saas-api_internal_auth.User'
    type: object
  cmd_api-gateway.resetPasswordResp:
    properties:
      message:
        type: string
    type: object
...
  github_com_r0lm0_go-saas-api_internal_auth.ForgotPasswordRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  github_com_r0lm0_go-saas-api_internal_auth.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  github_com_r0lm0_go-saas-api_internal_auth.RefreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  github_com_r0lm0_go-saas-api_internal_auth.RegisterRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - name
    - password
    type: object
  github_com_r0lm0_go-saas-api_internal_auth.ResetPasswordRequest:
    properties:
      new_password:
        minLength: 8
        type: string
      token:
        type: string
    required:
    - new_password
    - token
    type: object
  github_com_r0lm0_go-saas-api_internal_auth.TokenPair:
    properties:
      access_token:
        type: string
      expires_in:
        description: seconds
        type: integer
      refresh_token:
        type: string
    type: object
  github_com_r0lm0_go-saas-api_internal_auth.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
      oauth_provider:
        type: string
      oauth_subject:
        type: string
      role:
        type: string
      tenant_id:
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/auth/callback:
    post:
      consumes:
      - application/json
      parameters:
      - description: OAuth code and state
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cmd_api-gateway.loginResp'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Unified OAuth callback (r3-chat)
      tags:
      - auth
  /api/v1/auth/forgot-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Email address
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_r0lm0_go-saas-api_internal_auth.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cmd_api-gateway.forgotPasswordResp'
      summary: Request password reset email
      tags:
      - auth
...
  /api/v1/auth/github/callback:
    get:
      parameters:
      - description: Authorization code
        in: query
        name: code
        required: true
        type: string
      - description: OAuth state
        in: query
        name: state
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cmd_api-gateway.loginResp'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: GitHub OAuth callback
      tags:
      - auth
...
  /api/v1/auth/google/callback:
    get:
      parameters:
      - description: Authorization code
        in: query
        name: code
        required: true
        type: string
      - description: OAuth state
        in: query
        name: state
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cmd_api-gateway.loginResp'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Google OAuth callback
      tags:
      - auth
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_r0lm0_go-saas-api_internal_auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cmd_api-gateway.loginResp'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login with email and password
      tags:
      - auth
  /api/v1/auth/logout:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_r0lm0_go-saas-api_internal_auth.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cmd_api-gateway.logoutResp'
      summary: Logout and invalidate refresh token
      tags:
      - auth
...
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_r0lm0_go-saas-api_internal_auth.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cmd_api-gateway.refreshResp'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Refresh access token
      tags:
      - auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Registration info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_r0lm0_go-saas-api_internal_auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/cmd_api-gateway.registerResp'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - auth
  /api/v1/auth/reset-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Reset token and new password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_r0lm0_go-saas-api_internal_auth.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/cmd_api-gateway.resetPasswordResp'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Reset password with token
      tags:
      - auth
...
swagger: "2.0"